refactor(editor): extract selection move projection
Move visual-selection translation, command-move origins, instance Smart Snap projection, electrical route attachment, and final move commits into a selection-domain controller. useSelectionInteraction no longer depends on App implementations for its core move geometry and transaction semantics.

Validation: pnpm gate:plan -- --path apps/editor/src/features/selection/selection-move-controller.ts --path apps/editor/src/app/App.tsx; pnpm typecheck; pnpm test:local apps/editor/src/features/selection/use-selection-interaction.test.ts apps/editor/src/features/selection/selection-move-plan.test.ts apps/editor/src/app/App.test.tsx; pnpm test:e2e:local apps/editor/e2e/manual-editor.spec.ts --grep command move owns rotate and commits pose plus translation atomically|command move turns a component while locally stretching its boundary wire|moves an isolated free wire as one route|previews a connected Wire while its Instance moves|moves internal wiring with a selected group and copies the routed subgraph|selects and moves multiple instances while viewport gestures stay transient; pnpm ci:static; git diff --check

Test-Impact: no-test-change — existing selection planning, App, and six browser move contracts cover unchanged projection and commit behavior
